Add Score and gyeol button
